Fix typos, repeated words, align spelling to LDP
This is follow-up commit to
and concerns all the .c-, .h, .sh, .pod, .txt, .md
and Configure files of the repository.

The change standardizes the spelling mostly according
the rules of the Linux Documentation Project aside
of typos and repeated words. The changes affect the
source code comments and static variable names.
I have not changed the 2 tech-jargon cases
   alloced -> allocated
  "config file" -> "configuration file"
which would have larger impact.